requirement.txt
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
schedule==1.2.0
tweepy==4.14.0
